Add models, resources, factories, and API routes for listings and related entities
- Created models for Address, Booking, Category, Feature, HouseType, Image, Listing, ListingAvailableDate, ListingCategory, ListingFeature, ListingReview, Notification, Setting, Status, Transaction, UserListingFavorite, and UserVerification.
- Implemented corresponding resource classes for API responses.
- Added factories for all newly created models to facilitate testing and seeding.
- Established API routes for user authentication and retrieval.

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('first_name');
             $table->string('last_name');
-            $table->bigInteger('wallet_balance');
+            $table->float('wallet_balance');
             $table->string('avatar')->nullable();
             $table->string('email');
             $table->boolean('email_verified')->default(false);
